refactor: decompose testdata classes
