+++ src/backend/SportDataService.Tests/UnitTests/UseCases/TournamentResult/GetTournamentResultByResultIdUseCaseTests.cs
@@ -28,8 +28,8 @@ public class GetTournamentResultByResultIdUseCaseTests
         // Arrange
         var validTournamentResultId = "10012345";
         var ct = CancellationToken.None;
-        var testTournamentResult = UseCasesTestData.CreateTestTournamentResultsWithMetadata(1).First();
-        var testTournamentResultDto = UseCasesTestData.CreateTestTournamentResultDtos(1).First();
+        var testTournamentResult = TournamentResultUseCasesTestData.CreateTestTournamentResultsWithMetadata(1).First();
+        var testTournamentResultDto = TournamentResultUseCasesTestData.CreateTestTournamentResultDtos(1).First();
 
         _tournamentResultRepositoryMock
             .Setup(x => x.GetTournamentResultByTournamentResultIdAsync(validTournamentResultId, ct))
@@ -129,7 +129,7 @@ public class GetTournamentResultByResultIdUseCaseTests
         // Arrange
         var validTournamentResultId = "10012345";
         var ct = CancellationToken.None;
-        var testTournamentResult = UseCasesTestData.CreateTestTournamentResultsWithMetadata(1).First();
+        var testTournamentResult = TournamentResultUseCasesTestData.CreateTestTournamentResultsWithMetadata(1).First();
         var expectedException = new AutoMapperMappingException("Mapping failed");
 
         _tournamentResultRepositoryMock
